data upload for properties, city, and dataset tables is functioning
