--- server/modules/store.data.js
@@ -1,3 +1,0 @@
-
-
-module.exports = storage;
