+++ server/public/scripts/controllers/data.upload.controller.js
@@ -2,15 +2,9 @@ myApp.controller('DataUploadController', function ($location, DataService, UserS
     var vm = this;
     var ds = DataService
     vm.uploader = UploadService.uploader;
-    vm.uploaderOptions = {
-        url: '/data/csv/',
-        onSuccess: function (response, status, headers) {
-            console.log('data uploaded');
-        }
-    };
+    vm.uploaderOptions = ds.uploaderOptions;
 
-    vm.dataUpload = function () {
-        console.log('uploader options', vm.uploaderOptions);
-        vm.uploader.uploadItem(0);
+    vm.dataUpload = function (options) {
+        vm.uploader.uploadAll();
     };
 });
