+++ server/routes/email.router.js
@@ -1,11 +1,7 @@
 var express = require('express');
 var router = express.Router();
-var csv = require('fast-csv');
-var fs = require('fs');
 var pool = require('../modules/pool');
-
-// process lines of email CSV data to produce generated email addresses. Returns an array of data lines
-var processLine = require('../modules/address.generator');
+var processContactCSV = require('../modules/store.email.csv');
 
 var uploadedData = [];
 
@@ -254,86 +250,5 @@ function createBatch(dataInfo) {
     }); //end of pool
   });
 }
-
-
-function processContactCSV(dataInfo) {
-  return new Promise((resolve, reject) => {
-    fs.exists(dataInfo.path, function (exists) {
-      if (exists) {
-        var stream = fs.createReadStream(dataInfo.path);
-        csv.fromStream(stream, {
-          headers: true
-          // use the array below instead of 'true' if you wish to 
-          // omit headers from the CSV instead of the headers being
-          // in the first line of the CSV
-          //
-          // [
-          //     'first',
-          //     'last',
-          //     'title',
-          //     'company',
-          //     'domain',
-          //     'building',
-          //     'market',
-          //     'email'
-          // ]
-        })
-          .on("data", function (data) {
-            // generates email addresses from names and company domains and pushes the results to uploadedData
-            processLine(data).forEach((line) => {
-              dataInfo.uploadedData.push(line);
-            });
-          })
-          .on("end", function (data) {
-            resolve(storeEmailCSV(dataInfo));
-          });
-      }
-    });
-  });
-}
-
-
-// store email CSV in database
-function storeEmailCSV(dataInfo) {
-  return new Promise((resolve, reject) => {
-    let success = true;
-    let data = dataInfo.uploadedData;
-    let user = dataInfo.user;
-    let batchId = dataInfo.batchId;
-    console.log('attempting to store Email CSV');
-    data.forEach((contact) => {
-      pool.connect(function (errorConnecting, db, done) {
-        if (errorConnecting) {
-          console.log('Error connecting', errorConnecting);
-          reject(errorConnecting);
-        } else {
-          var queryText = 'INSERT INTO "emails" ("first","last","title","company","domain","building","market","email","office_id","batch_id") VALUES($1,$2,$3,$4,$5,$6,$7,$8,$9,$10);';
-          db.query(queryText, [
-            contact.first,
-            contact.last,
-            contact.title,
-            contact.company,
-            contact.domain,
-            contact.building,
-            contact.market,
-            contact.email,
-            user.o_id,
-            batchId
-          ], function (errorMakingQuery, result) {
-            done();
-            if (errorMakingQuery) {
-              console.log('error making email entry query', errorMakingQuery);
-              reject(errorMakingQuery);
-            } else {
-
-            }
-          });
-        }
-      }); //end of pool
-    });
-    resolve(batchId);
-  });
-}
-
 
 module.exports = router;
